Split instrumentation into client and consumer parts
