+++ lib/traceview/inst/bunny.rb
@@ -140,54 +140,11 @@ module TraceView
         end
       end
     end
-
-    module BunnyConsumer
-      def self.included(klass)
-        ::TraceView::Util.method_alias(klass, :call, ::Bunny::Consumer)
-      end
-
-      def collect_consumer_kvs
-        begin
-          kvs = {}
-          kvs[:Spec] = :job
-          kvs[:Flavor] = :rabbitmq
-          kvs[:RemoteHost]  = @channel.connection.host
-          kvs[:RemotePort]  = @channel.connection.port.to_i
-          kvs[:VirtualHost] = @channel.connection.vhost
-
-          if @queue.respond_to?(:name)
-            kvs[:Queue] = @queue.name
-          else
-            kvs[:Queue] = @queue
-          end
-
-          if TV::Config[:bunnyconsumer][:log_args] && @arguments
-            kvs[:Args] = @arguments.to_s
-          end
-
-          report_kvs[:Backtrace] = TV::API.backtrace if TV::Config[:bunnyconsumer][:collect_backtraces]
-
-          kvs
-        rescue => e
-          TraceView.logger.debug "[traceview/debug] #{__method__}:#{File.basename(__FILE__)}:#{__LINE__}: #{e.message}" if TraceView::Config[:verbose]
-        ensure
-          return kvs
-        end
-      end
-
-      def call_with_traceview(*args)
-        result = TraceView::API.start_trace('rabbitmq-consumer', nil, collect_consumer_kvs) do
-          call_without_traceview(*args)
-        end
-        result[0]
-      end
-    end
   end
 end
 
-if TraceView::Config[:bunny][:enabled] && defined?(::Bunny)
-  ::TraceView.logger.info '[traceview/loading] Instrumenting bunny' if TraceView::Config[:verbose]
+if TraceView::Config[:bunnyclient][:enabled] && defined?(::Bunny)
+  ::TraceView.logger.info '[traceview/loading] Instrumenting bunny client' if TraceView::Config[:verbose]
   ::TraceView::Util.send_include(::Bunny::Exchange, ::TraceView::Inst::BunnyExchange)
   ::TraceView::Util.send_include(::Bunny::Channel, ::TraceView::Inst::BunnyChannel)
-  ::TraceView::Util.send_include(::Bunny::Consumer, ::TraceView::Inst::BunnyConsumer)
 end
